Bug 1350377 - Implement `History.fetch` and accordingly change `promisePlaceInfo`, r=mak
`fetch` was written from scratch to query the database and get information for a
single page to replace `getPlacesInfo`. A parameter for visits was also added to
the signature (default by false) since it's feasible to get the visits for a single
page.

`promisePlaceInfo` was marked deprecated and turned into a wrapper for fetch.

MozReview-Commit-ID: 9ZYHOmrpCHg

UltraBlame original commit: 1eac66ddfb7332e399dca773f65947dee1cb8bad