Bug 989373, part 3 - |function| is always null in ObjectForLocation. r=kmag
With the changes in part 1, the local variable |function| is always
null in ObjectForLocation, which lets me remove a lot of dead
code. Additionally, in the second half of the function we know that
|script| is always non-null, because we return with a failure before
then if it isn't null.

In addition, with these changes WriteCachedFunction is no longer
called, so I removed it.

MozReview-Commit-ID: 60hEPi8S3H4

UltraBlame original commit: d9f3972b7d978da13ce781ae42a9140f97a50c9c