Bug 1366637 - Extend expiration on MEDIA_OGG_LOADED_IS_CHAINED telemetry. r=francois
In order to inform our decision as to when to remove support for
ogg chaining, we should restore the collection of the telemetry
that reports when it's used.

I picked a longer term on the probe, as I expect it will take a
while for usage to drop in the wild.

MozReview-Commit-ID: Kk8kYJw9Opv

UltraBlame original commit: 5f533c9599661f22dcf2e182bcc96afd6b6a653c