Bug 1363182 - Add a "send to device" subview to the page action menu. r=eoger,mikedeboer
Add a Send to Device subview to the page action panel.  When the page isn't
sendable, disable the Send to Device menu item.  When the user doesn't have any
devices, show a menu item that opens the Firefox Account preferences pane.

Generalize gSync.populateSendTabToDevicesMenu() so that it can be used to
populate any kind of container, not only a menupopup with menuitems.

Add an SVG that shows a phone and an SVG that shows a desktop.

MozReview-Commit-ID: EZQKAEAr08q

UltraBlame original commit: 15688bfe56c9899bf050e69d3d0cf999e084e297