Bug 1367568 part 2: Rewrite nsChangeHint_ReflowHintsForBSizeChange in terms of nsChangeHint_AllReflowHints. r=bz
This patch doesn't affect behavior at all.  It's simply adjusting the way we
express a bitmask, without modifying the actual value of that bitmask.  The
intent of this change is to make this bitmask's definition as similar as
possible to the definition for its ISize (inline-axis) analog, so that it's
easy to compare the two.

MozReview-Commit-ID: FUJnBQm8hmJ

UltraBlame original commit: 20d8dd705b0ebe84de47e0ffb4534b6d645feaf0