Bug 1364588 - part 2 - move Stylo package installation; r=rillian
Now that we're installing Stylo packages through `mach`, we need to have
`mach` available, which is only true after we've checked for a clone of
the tree.  We move the block performing the installation in this commit
and then fixup resulting problems in future commits.

UltraBlame original commit: c37162881c6e5ccfbea82e5377ba2fa859c2c9b8