Bug 1366823 - Use DevToolsShim to call Scratchpad APIs from SessionStore.jsm;r=mikedeboer
DevTools are moving out of mozilla-central and m-c code should only rely on the
DevToolsShim that will remain available in m-c after the migration.
This changeset updates SessionStore.jsm to call the DevToolsShim instead of the
scratchpadmanager.

The actual underlying implementation is the same when devtools are installed and
no-ops when they are not.

MozReview-Commit-ID: Fv22TtjJiDw

UltraBlame original commit: f39ea1fb1be510633bca4cfe7cf57b9506831e39