Bug 1362165: P4. Don't assume the duration found when reading metadata will be the final one. r=gerald
When endOfStream() is called, the duration is changed to that of the last buffered frame's end time.

The ended event will only be fired if currentTime is equal to the duration. So if the duration has changed following a call to endOfStream, then playback won't be considered as ended.

MozReview-Commit-ID: DBu3LorwFTI

UltraBlame original commit: 5334cc6d0ea9b47325f4d6fcf4dab59307d12d55