Bug 1365204 - Add update badge histogram r=bsmedberg,rstrong
Currently we can't differentiate between when a badge is shown
and a doorhanger is shown. This creates an additional problem
where if the badge progresses into a doorhanger after a window of
time has passed, it registers as two notifications shown, when
logically it is one. This splits out badges and doorhangers to
remedy that.

MozReview-Commit-ID: CTTaWDG1tah

UltraBlame original commit: 860549ef693bc8237256435f59a5986da63d97ef