Bug 1367980 - don't release decoder in complete state on Fennec. r=jwwang
On Android, the life cycle of graphic buffer is equal to Android's codec, and
it would be released immediately if we shutdown the decoder.

Since the buffer has been release, the memory we access is totally meaningless.
The result is we would render the black image on the screen.

In addition, I suspect this issue might be one of the root cause which results
in lots of timeout for media mochitest, so I would like to land this patch first,
and then do the follow-up in bug1367983.

MozReview-Commit-ID: 5eZph6MItZs

UltraBlame original commit: 04c6a723dddcdb76456b8c58a90445d6ea490fe7