Bug 1363678 - Allow to pass a custom logLimit to configureStore. r=bgrins
The idea is to allow tests to set their own logLimit without
having to use Services.setIntPref.
To do so, we also need to not retrieve the logLimit from Services
in the messages reducer. Since we already have the logLimit in the
store, we can retrieve it directly from the store by passing the
prefs state to the messages reducer.
We take this as an opportunity to revisit the createRootReducer function
to be future proof: if any new reducer is added, they should just work
as if we were using combineReducers.

With those changes, we can set a lower logLimit in some tests that used
to have a defined timeout since they were taking too much time.

MozReview-Commit-ID: 7j80XoKkJ1y

UltraBlame original commit: 39d218ebd0d6c3d0e677d22b860485aabc77c592