Bug 1369382: Don't make hard assertion that the queued sample is a keyframe. r=kaku
When seeking, we go to the first IDR prior the seek target. However, if the MP4 sample table is invalid the frame returned may not be an IDR.

It is an error for that frame to not be a keyframe, so we do want to have way to detect them through logging. Change assertion to a soft assertion for that purpose.

MozReview-Commit-ID: EMgwRo1mYMp

UltraBlame original commit: 5fe1447e3756f8db5b698f1327c2231316a705aa