Bug 1337007 - Clear umask while updating on OSX r=rstrong,spohl
On OSX, we want the updater to behave the same as when a user
drags a .dmg to the /Applications directory. When the user does
this, OSX does not follow their umask settings, assuming the user
wants to install the application for all users. The updater
should behave similarly.

MozReview-Commit-ID: JXFI9mW9erL

UltraBlame original commit: b055ba326f1f2930f1ca12e1927f21521fa07b80