Bug 1341404 - add linux64 noopt debug builds; r=dustin,mshal
A number of developers find it convenient to build with
--disable-optimize --enable-debug for an improved debugging experience.
We don't currently have a configuration in CI that ensures this
combination of options works, so various changes break builds with this
configuration every so often.  We should test such configurations to
ensure they build to provide a smooth experience for developers.

UltraBlame original commit: 6e538d5add9adf7af315b8fede3c695d23491663