Bug 1353168 - Add a crash annotation to distinguish between web, file and extension content processes; r=bsmedberg
This adds the RemoteType annotation to a content crash report so that we can
distinguish between content processes that crashed while running remote, local
or extension code. The annotation is passed along the others to Socorro by the
crashreporter and is also whitelisted for inclusion in the crash ping.

MozReview-Commit-ID: 4avo0IWfMGf

UltraBlame original commit: 9d26051a28ea4bffdc654357855f50c9a685fbcc