Bug 1369866 - Use TOOLCHAIN_PREFIX when running dyldinfo; r=ted
The test in old-configure.in for MOZ_REPLACE_MALLOC_LINKAGE runs
dyldinfo, which is a native tool on OSX. We have it in the cross
toolchain as well, but as x86_64-apple-darwin11-dyldinfo. We should use
the TOOLCHAIN_PREFIX here to make sure we get the same result as native
builds.

MozReview-Commit-ID: 3jyzpaM8ZGy

UltraBlame original commit: bfb69fd3ed936ded13e9b4210ab158b823f99c49