Bug 1371100 - Check userContextId in the content process instead of in the chrome process, r=smaug
This is necessary because we asynchronously send the updated UserContextId back
from the child to the parent process, and we would occasionally lose the race on
infrastructure causing us to see the non-updated version.

With this change we explicitly ask the content process, which is not racey.

MozReview-Commit-ID: LODNi1YmBI8

UltraBlame original commit: 66a83f17d67b97e32e65600fb636f05336114bc4