Bug 1370986 - switch to photonpanelmultiview for photon, disable off-nightly, r=mikedeboer
This #ifdefs out the multiview for non-photon-theme, and checks for it being
present in various bits of JS that interact with it. As a result, this will
'fix' the issues in this bug and in bug 1370967 for 55 as it moves off
Nightly. By switching to the photonpanelmultiview, we get proper anchoring
and slightly improved styling (bug 1354086 covers the rest of that), even
where this code *is* enabled. bug 1370967 will still need fixing in
the photonpanelmultiview / webextensions.

MozReview-Commit-ID: 6x4HmyvxeRP

UltraBlame original commit: 71da3a25b5720039ef7a0c84e0b49a35c86cae0a