Bug 1360525 - Separate repackage out into subcommands; r=chmanchester
With multiple types of repackaging, it will be simpler to manage the
various commands if they are organized as sub-commands. This way the
different repackage types can take varying arguments, and we don't have
to guess on the repackage type based on the output filename.

MozReview-Commit-ID: BknRPAwFG5H

UltraBlame original commit: 91b810a4054786b9eb6671b036435ab2f9eb8069