Bug 1370486 - Add mocha tests for messages store cleaning. r=Honza
In the process of adding those tests, I realized the cleanup was
not effective. There were some bugs for Immutable List (we can't
use delete in a withMutation callback), and in object/arrays (we
we're using Array.includes with another array as a parameter).
These were fixed, and the new tests are now passing.

MozReview-Commit-ID: NrKjLuu25

UltraBlame original commit: ea7907038c5f5d77803ef1d85f59918836cee54f