Bug 1371923 - Only check safebrowsing files when they are in preference in test_safe_browsing_initial_download.py. r=francois
Ensure that we only wait for the providers that are enabled (i.e. have active lists) to avoid waiting forever
and timing out.

In order to check whether or not the google and google4 providers are enabled, we look for lists that start
with "goog" in both the V2 files and the V4 files. We use 'goog' here instead of 'goog-' because there are
both 'goog-xxx-xxx' and 'googpub-xxx-xxx' tables

MozReview-Commit-ID: 72JWZRs3Qxd

UltraBlame original commit: 92846886fda2655c3dbd8db935d51cf41a9bcc33