Bug 1374127 (part 3) - Reorder macros in GeckoProfiler.h. r=mstange.
It's bugged me for some time that the comments explaining what all the
PROFILER_* macros do are on the empty definitions that are used when the
profiler is disabled. This patch switches the order.

UltraBlame original commit: cf6c6bbc2816a5d0966a2b6ae9d9c19e5fe0fb97