Bug 1286802 - Part 4: Implemention of including heap regions from the crashing context. r=ted
This patch is the core part of including heap regions using the registers of
the crashing context to be used by in-process and out-of-process crash
generations. It works as follows:

When a crash happens, the ExceptionHandler or CrashGenerationServer uses the register
values in the crash contex as candidates for heap regions. If a value "looks like"
a pointer to some data on the heap, we register a region around the value using a
preallocated AppMemory instance. Then in the "MemoryCallback" type of MinidumpCallback,
we register the memory regions for inclusion in the minidump.

MozReview-Commit-ID: BEik3lRJJkj

UltraBlame original commit: e76c0ff35226451fbbb9cd66d02a1d5c6264986c