Bug 1307954 - Add integration test for clicking on network event message url; r=Honza.
This ensures that clicking on a message open the netmonitor panel, with the network
item clicked being selected in the netmonitor panel.
The loadDocument was edited to accept a URL instead of relying on the existence of
a TEST_PATH variable. Callsites where thus modified.

MozReview-Commit-ID: 3dbX1Myxirz

UltraBlame original commit: f0bd396cd081e240067189e80ea52d61716487ce