Bug 1374824 - Fix --enable-stylo value processing; r=froydnj
The argument passed to the function is a PositiveOptionValue,
which represents its option values as a tuple. The __eq__ for
this type first compares type() of the operands. Since the
previous code compared a PositiveOptionValue to a string
literal, this always failed. There's possibly room to
improve the behavior of PositiveOptionValue. But for now,
let's rewrite stylo_config() so it works.

MozReview-Commit-ID: B4vkYwCDHrb

UltraBlame original commit: 1b500030019011e9ffbf26b9cb5f1990423a4d6e