Bug 1107364 - remove findNodeByDetails and maintain a map to fetch a node by details speedily, r=mak
This patch removes the interface/header/implementation.

This adds a Map to PlacesTreeView to maintain the relation between node
details and the nodes, based on changes in `this._rows` or on
node details changed events. The Map exists from node details
to nodes (not rows).

MozReview-Commit-ID: EUNiXNIB5rN

UltraBlame original commit: 87b2b354ed01650863638342d58a8d2496f4797a