Bug 1369670 Part 1: Resolve to the real path names when comparing the profile paths in test_capabilities.py. r=whimboo
Unlike python 3, python 2 doesn't have a proper implementation of realpath or
samefile for Windows. However there is a function convert_path in FixTk.py that
was added to python 2 to fix an issue with tcl installations and symlinks, so
we use that as this is just for testing.

UltraBlame original commit: 9120668e5a82faeaa98df34616355620ea774c8e