Bug 1376392 - do a more thorough check for git checkouts of mozilla-central in bootstrap; r=rillian
When determining the path of a possible mozilla-central checkout,
bootstrap currently considers the existence of a `.git` directory as
sufficient proof that the containing directory is a git checkout.
Unfortunately, if you happen to execute standalone bootstrap from a git
checkout of something else, you're gonna have a bad time.

To prevent this, check for the existence of a moz.configure file.  This
is not an ideal proof, but it is better than what we currently have.

UltraBlame original commit: 0dd4dc37770db5952c34df4899578fa8589b586c